Songs Per Year
This is the plot for number of songs per year plotted on plot.ly.Since
we have used approximately less than half of the dataset, we have a
total of 256139 songs.

This has been plotted against the year in which they are released.
The data file for plotly is data_for_plotly.csv